+++ engine/src/atlas/renderer/Color.cpp
@@ -8,10 +8,10 @@ namespace Atlas {
         }
 
         color = glm::vec4(
-            std::stoi(hex.substr(0, 2), nullptr, 16) / 255.0f,
-            std::stoi(hex.substr(2, 2), nullptr, 16) / 255.0f,
-            std::stoi(hex.substr(4, 2), nullptr, 16) / 255.0f,
-            std::stoi(hex.substr(6, 2), nullptr, 16) / 255.0f
+            static_cast<float>(std::stoi(hex.substr(0, 2), nullptr, 16)) / 255.0f,
+            static_cast<float>(std::stoi(hex.substr(2, 2), nullptr, 16)) / 255.0f,
+            static_cast<float>(std::stoi(hex.substr(4, 2), nullptr, 16)) / 255.0f,
+            static_cast<float>(std::stoi(hex.substr(6, 2), nullptr, 16)) / 255.0f
         );
     }
 }
