+++ engine/src/atlas/renderer/Sampler.cpp
@@ -90,7 +90,7 @@ namespace Atlas {
         }
 
         transitionImageLayout(VK_IMAGE_LAYOUT_UNDEFINED, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL);
-        copyBufferToImage(stagingBuffer.get(), textureImage, static_cast<uint32_t>(width), static_cast<uint32_t>(height));
+        copyBufferToImage(stagingBuffer.get(), textureImage, width, height);
         transitionImageLayout(VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL);
     }
